Former-commit-id: 2846b86fc40f47220396aeeedb46a401a94429da

+++ tests/loss_map_output_unittest.py
@@ -89,11 +89,6 @@ SAMPLE_LOSS_MAP_NON_DET_DATA = [
 
 GML_ID_KEY = '{%s}id' % xml.GML_NS
 
-DEFAULT_METADATA = risk_output.LossMapXMLWriter.DEFAULT_METADATA
-
-DEFAULT_NON_DET_METADATA = \
-        risk_output.LossMapNonScenarioXMLWriter.DEFAULT_METADATA
-
 LOSS_MAP_NODE_ATTRS = ('endBranchLabel', 'lossCategory', 'unit')
 LOSS_MAP_NON_DET_NODE_ATTRS = ('endBranchLabel', 'lossCategory', 'unit',
     'timeSpan', 'poE')
@@ -104,36 +99,34 @@ class LossMapOutputTestCase(unittest.TestCase):
     engine is valid against schema, as well as correct given the inputs."""
 
     def setUp(self):
-        self.xml_writer = \
-            risk_output.LossMapXMLWriter(TEST_LOSS_MAP_XML_OUTPUT_PATH)
-        self.xml_non_det_writer = \
-            risk_output.LossMapNonScenarioXMLWriter(
-                    TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH)
+        self.xml_writer = risk_output.LossMapXMLWriter(
+            TEST_LOSS_MAP_XML_OUTPUT_PATH)
+        self.xml_non_det_writer = risk_output.LossMapNonScenarioXMLWriter(
+            TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH)
 
     def tearDown(self):
         self.xml_writer = None
         self.xml_non_det_writer = None
-        os.remove(TEST_LOSS_MAP_XML_OUTPUT_PATH)
-        #os.remove(TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH)
+        if os.path.exists(TEST_LOSS_MAP_XML_OUTPUT_PATH):
+            os.remove(TEST_LOSS_MAP_XML_OUTPUT_PATH)
 
     def test_loss_map_output_writes_and_validates(self):
-        xml_writer = \
-            risk_output.LossMapXMLWriter(TEST_LOSS_MAP_XML_OUTPUT_PATH)
+        xml_writer = risk_output.LossMapXMLWriter(
+            TEST_LOSS_MAP_XML_OUTPUT_PATH)
         xml_writer.serialize(SAMPLE_LOSS_MAP_DATA)
         self.assertTrue(
             xml.validates_against_xml_schema(TEST_LOSS_MAP_XML_OUTPUT_PATH,
             NRML_SCHEMA_PATH),
-            "NRML instance file %s does not validate against schema" % \
+            "NRML instance file %s does not validate against schema" %
             TEST_LOSS_MAP_XML_OUTPUT_PATH)
 
-        xml_non_det_writer = \
-            risk_output.LossMapNonScenarioXMLWriter(
-                    TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH)
+        xml_non_det_writer = risk_output.LossMapNonScenarioXMLWriter(
+            TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH)
         xml_non_det_writer.serialize(SAMPLE_LOSS_MAP_NON_DET_DATA)
         self.assertTrue(
             xml.validates_against_xml_schema(
                 TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH, NRML_SCHEMA_PATH),
-            "NRML instance file %s does not validate against schema" % \
+            "NRML instance file %s does not validate against schema" %
             TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH)
 
     def test_write_metadata(self):
@@ -149,7 +142,7 @@ class LossMapOutputTestCase(unittest.TestCase):
             for key, node in (
                 ('nrmlID', xml_writer.root_node),
                 ('riskResultID', xml_writer.risk_result_node),
-                ('lossMapID', xml_writer.loss_map_node)):
+                ('lossMapID', xml_writer.map_container)):
 
                 self.assertEqual(
                     loss_map_data[key],
@@ -157,15 +150,15 @@ class LossMapOutputTestCase(unittest.TestCase):
 
             if isinstance(xml_writer,
                     risk_output.LossMapNonScenarioXMLWriter):
-                loss_map_node_attrs = LOSS_MAP_NON_DET_NODE_ATTRS
+                map_container_attrs = LOSS_MAP_NON_DET_NODE_ATTRS
             else:
-                loss_map_node_attrs = LOSS_MAP_NODE_ATTRS
+                map_container_attrs = LOSS_MAP_NODE_ATTRS
 
             # Verify the <lossMap> attributes
-            for key in loss_map_node_attrs:
+            for key in map_container_attrs:
                 self.assertEqual(
                     loss_map_data[key],
-                    xml_writer.loss_map_node.attrib[key])
+                    xml_writer.map_container.attrib[key])
 
     def test_write_metadata_with_some_defaults(self):
         """
@@ -185,16 +178,16 @@ class LossMapOutputTestCase(unittest.TestCase):
             xml_writer.write_metadata(partial_meta)
 
             self.assertEqual(
-                DEFAULT_METADATA['nrmlID'],
+                'undefined',
                 xml_writer.root_node.attrib[GML_ID_KEY])
 
             self.assertEqual(
-                DEFAULT_METADATA['riskResultID'],
+                'undefined',
                 xml_writer.risk_result_node.attrib[GML_ID_KEY])
 
             self.assertEqual(
-                DEFAULT_METADATA['unit'],
-                xml_writer.loss_map_node.attrib['unit'])
+                'undefined',
+                xml_writer.map_container.attrib['unit'])
 
     def test_serialize_with_no_meta(self):
         """
@@ -207,30 +200,28 @@ class LossMapOutputTestCase(unittest.TestCase):
 
             if isinstance(xml_writer,
                     risk_output.LossMapNonScenarioXMLWriter):
-                loss_map_node_attrs = LOSS_MAP_NON_DET_NODE_ATTRS
+                map_container_attrs = LOSS_MAP_NON_DET_NODE_ATTRS
                 # everything but metadata
                 test_data = SAMPLE_LOSS_MAP_NON_DET_DATA[1:]
-                default_metadata = DEFAULT_NON_DET_METADATA
             else:
-                loss_map_node_attrs = LOSS_MAP_NODE_ATTRS
+                map_container_attrs = LOSS_MAP_NODE_ATTRS
                 # everything but metadata
                 test_data = SAMPLE_LOSS_MAP_DATA[1:]
-                default_metadata = DEFAULT_METADATA
 
             xml_writer.serialize(test_data)
 
             for key, node, in (
                 ('nrmlID', xml_writer.root_node),
                 ('riskResultID', xml_writer.risk_result_node),
-                ('lossMapID', xml_writer.loss_map_node)):
+                ('lossMapID', xml_writer.map_container)):
                 self.assertEqual(
-                    default_metadata[key],
+                    'undefined',
                     node.attrib[GML_ID_KEY])
 
-            for key in loss_map_node_attrs:
+            for key in map_container_attrs:
                 self.assertEqual(
-                    default_metadata[key],
-                    xml_writer.loss_map_node.attrib[key])
+                    'undefined',
+                    xml_writer.map_container.attrib[key])
 
     def test_loss_map_xml_content_is_correct(self):
         """
@@ -269,17 +260,15 @@ class LossMapOutputTestCase(unittest.TestCase):
                 (self.xml_non_det_writer, SAMPLE_LOSS_MAP_NON_DET_DATA)):
             xml_writer.serialize(loss_map_data)
 
-            if isinstance(xml_writer,
-                    risk_output.LossMapNonScenarioXMLWriter):
+            if isinstance(xml_writer, risk_output.LossMapNonScenarioXMLWriter):
                 expected_loss_map = EXPECTED_TEST_NON_DET_LOSS_MAP
-                test_loss_map_xml_output_path = \
-                        TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH
+                lossmap_xml_out_path = TEST_NON_DET_LOSS_MAP_XML_OUTPUT_PATH
             else:
                 expected_loss_map = EXPECTED_TEST_LOSS_MAP
-                test_loss_map_xml_output_path = TEST_LOSS_MAP_XML_OUTPUT_PATH
+                lossmap_xml_out_path = TEST_LOSS_MAP_XML_OUTPUT_PATH
 
             expected_elems = get_xml_elems_list(expected_loss_map)
-            actual_elems = get_xml_elems_list(test_loss_map_xml_output_path)
+            actual_elems = get_xml_elems_list(lossmap_xml_out_path)
 
             # first, make sure out elem list from the expected data
             # is _not_ empty (otherwise, this would be an epic test fail)
